Additional refactoring and commenting
